+++ NEWS.md
@@ -4,6 +4,7 @@
 * Bugfixes
     * Fix an error where a pointer was freed even though it was not allocated
       (issue #356; @dschlaep).
+    * Fix memory leak in test of `SW_VPD_construct()` (issue #205; @dschlaep).
 
 # SOILWAT2 v7.0.0
 * This version produces nearly identical simulation output